cpu/sam0_common/periph_gpio_ll: fix gpio_query_conf()
For the other MCUs, we take the input register state instead of the
output register state when the pin is configured as input. Let's do
the same here, as this is a lot more useful and intuitive.

(cherry picked from commit 0222b8c54cda20baec9a95d5357d18eb958efa1f)