pkg/lwip: fix adaption to API change of netdev_new_api
Since https://github.com/RIOT-OS/RIOT/pull/21012 a netdev in the new API
can return > 0 directly in netdev_driver_t::send() to indicate the
driver is naturally synchronous and has already completed the
transmission.

The adaption of lwIP to that API change contained a bug: It handled the
case after the thread is already blocked waiting for the signal that
is never going to arrive. This is now fixed.