examples/gnrc_border_router: static route for dhcp
The gnrc_border_router example uses by default the start_networking
script when used via `make term`. This script in turn uses the KEA
DHCPv6 server which does not automatically configures a downstream route
toward the delegated prefix. Hence, it is the responsibility of the
caller to setup the route themselves. This can be done with RIOT when
`STATIC_ROUTES` is set to `1`. (This will add a static link-local
address to the border router's upstream interface (`fe80::2`) and a
route via this address to the delegated prefix will be configured by the
start_networking script.)