tests/unittests: add hack for nanocoap
nanocoap depends on sock_udp, which in turn needs `sock_types.h` to be
available. When no network stack is in use, that sadly is not the case.
To work around the issue, we just manually add the include paths if
`gnrc_sock` is not used.

This fixes an issue with compilation of only the nanocoap unit tests
without also including the GNRC tests.

(cherry picked from commit e22e3573745647ff115f0c98df7161963b3192ec)