drivers/can: add board_candev_set_power() hook
This provides a default (weak) implementation of

    void board_candev_set_power(candev_t *dev, bool active);

that is a no-op, and adds calls to this function to each and every CAN
controller driver. The idea is that a board can provide this, if it
can power up and down the CAN transceivers, freeing applications from
having to add board specific hacks to enable CAN transceivers.

In addition, the following two helpers are added:

    can_t *candev2periph(candev_t *candev);
    candev_mcp2515_t *candev2mcp2515(candev_t *candev);

These safely convert a `candev_t` pointer back to the underlying type,
return `NULL` if the given `candev_t *` is not of the requested type.
This aids in implementing `board_candev_set_power()`, as it provides
a board with the means to figure out for which CAN controller the
CAN transceiver needs to be enabled/disabled for.