sys/event/periodic: fix event count footgun
As the event count code is written, I believe that it should not work
according to the rules of operator precedence for c. Perhaps some
compiler optimization is covering the issue up though.

Regardless, swapping the left and right hand sides of the && operator
does create a problem and the statement just looks like a potential
footgun. Further, when I did trigger the problem the counter goes
negative counting down. Eventually it should hit 0 and stop periodic
events which don't have a counter set on them. Therefore, this would be
a very hard to find bug.

This patch makes the code much more explicit to defend against future
bugs.