buildsystem: Enable -Wdocumentation for LLVM
This enables warnings for incorrect Doxygen comments when build with
LLVM. To my experience, LLVM provides a low false positive rate
(e.g. it does not complain about undocumented internal macros) and has
very helpful output.

This disable `-Wdocumentation-deprecated-sync` for now, as we do not
deprecate code with attributes yet.

The warnings are not set as errors, so that we can enable this without
breaking builds.