cpu/samd5x/periph_can: Make TXP configurable
Allow the board to configure whether the transmit pause feature of the
CAN controller should be used or not.

The default has been changed to enable the transmit pause feature. This
will slightyly decrease the maximum throughput by inserting a 2 bit time
pause between two subsequent frames to send. The benefit is that it will
ensure that nodes on the CAN bus with a lower priority (according to
their CAN ID) will not starve, effectively prevending the
"Babbling Idot Syndrome".