tests/unittests/test-core: improve test for clist_sort()
Increase the test coverage of the `clist_sort()` unit test to cover more
more edge cases (list with no items in it, list with one item in it).

In addition, the verification will not check if the result is actually
sorted, not only that the last element is the highest.