tests/bench/runtime_coreapis: reduce benchmark runs in the CI
On `native32` / `native64` in the CI, the benchmark results will be
unusable due to the background load of other build tasks anyway. There
is a risk that due to the high number of benchmark runs the test will
time out.

So to avoid the timeouts and because the results are unusable anyway,
the benchmark runs have been drastically reduced. This will still
allow the CI to confirm that the benchmarks do run successfully to
completion.