dist/tools: add a leading whitespace check for #pragma once
When adding the #pragma once after a comment block, many editors
will remain at the previous indentation level, adding a leading
whitespace to the #pragma once. This is invalid, but causes the
headerguards check to fail. Since it is a common issue, it
warrants a separate check with a proposed solution, just like
for other headerguard issues that are checked.