cpu/esp32/periph/hwrng: enable/disable bootloader RNG
If the ADC SAR is used, the Bootloader RNG must not be enabled before the random numbers are actually required. The reason is that the Bootloader RNG uses the noise of the ADC SAR reference voltage as a non-RF entropy source. The calibration of the ADC SAR does not work correctly in this case. Therefore, the Bootloader RNG is only enabled if random numbers are really required.