sys/embunit: fix return value on failure
When using `OUTPUT=COLORTEST` the return value of the unit test's
`main()` on failure was `0`, while it should be `1`. As a result,
running the unit tests from a script using

    $ OUTPUT=COLORTEXT make BOARD=native64 RIOT_TERMINAL=native -j32 flash term -C tests/unittests

would not catch a test failure.

This commit changes the behavior so that `0` is returned on success
and `1` on failure for both regular and colored text output.