ci: tune down static test to a warning
Murdock will run the static tests again after a rebase and will fail
hard.

There have been a few instances where static tests failed due
to e.g. a newer version of codespell finding more typos, or a newer
docker image needing to be tagged. Even PRs completely unrated to this
needed to be rebased on `master` to unblock the merge. If we tune down
the error to a warning, maintainers would be enabled to just merge
anyway without the rebase, relying on the merge commit to still work.
An accidental merge of failing static test would still be prevented by
Murdock.