dist/tools: add env. check to esptools/export.sh, enh. docs
The esptools/export.sh script has to be sourced, otherwise
the exported variables are not persistent in the environment,
rendering the script useless. The commit adds a check to the
script that it runs in the correct environment and enhances the
documentation to make it clearer that the "source" prefix command
is important.