dist/tools: various improvements of esptools scripts, enh. docs
The esptools/export.sh script has to be sourced, otherwise
the exported variables are not persistent in the environment,
rendering the script useless. The commit adds a check to the
script that it runs in the correct environment and enhances the
documentation to make it clearer that the "source" prefix command
is important.

The PR also introduces a version check for the PATH variable
to notify the user if an old version is present.