gnrc/netapi/netnotify: ACK notify event directly
Use `thread_flags` to acknowledge a received netnotify event as soon as
the associated event data was read. This avoids potentialy deadlocks
that could happen if the receiver's handling of the event involves an
exclusive lock on `netreg`.

Instead of using `msg_send_receive`, the sender now blocks on the
thread flag.

Add helper functions in `netnotify` that take care of parsing the
different event data types and directly setting the ACK flag.