dist/tools/vera++: allow comments to be overlong
The line length check is very noisy, especially in the Github web view
with annotations added to every overlong line. Reducing the number of
false positives there for is very valuable.

In comments, we often have overlong lines e.g. because an URL is longer
than 100 chars, because we add a larger markdown table, we add ASCII
art to visualize data structures, or because we add dot syntax that may
just need overlong lines.

Since vera++ does not support annotations with magic comments to disable
the length check just for a table / URL / graphic, this just disables
the length checks for all comments outright.