cpu/esp32/esp-idf/rmt: fix compilation together with Nimble and WiFi
The RMT driver module `esp_idf_rmt` uses GDMA on ESP32-S3, which in turn requires the activation and deactivation of flash encryption. Therefore, when the RMT driver module `esp_idf_rmt` is used, the function `gdma_disconnect` is used, which in turn requires the function `esp_flash_encryption_enabled`. However, if the NVS flash module `esp_idf_nvs_flash` is also used for some reason, e.g. because WiFi or BLE is used, the function `esp_flash_encryption_enabled` is already compiled in the module `esp_idf_nvs_flash` and must not be compiled again in the module `esp_idf_rmt`.