drivers/dose: implement netdev API async
This changes the behavior of the netdev implementation to match what
async drivers do. The current implementation is synchronous, but that
can be fixed once an async UART API is available.

In any case, we rather should not complicate testing of the network stack
by having network devices come in different flavors.