cpu/native: fix FPU/SSE register related context switching issues
When the `float_math` feature is not in used, this adds CFLAGS to
instruct the compiler to not make use of FPU/SSE registers, fixing
issues (random crashes, floating point exceptions, stack corruptions,
incorrect computations) due to incorrect back and restore of those
registers on context switching.

If `float_math` is used, the buggy behavior as before is produced.
But since `float_math` is not actually provided by the native CPU,
one would have to compile `CONTINUE_ON_EXPECTED_ERRORS=1` to actually
get into that situation.

Co-authored-by: Mihai Renea <mihai.renea@ml-pa.com>