posix/sockets: Prevent deadlock with multiple TCP sockets
When 2 posix TCP sockets in different threads are open and the first
one is in accept() the mutex of the socket pool was not unlocked
until data was received. When another thread calls a function which
needs the socket pool mutex there was a deadlock. Unlocking the mutex
after all operations on the socket pool are done, in particular BEFORE
sock_tcp_accept() solves this.