drivers/netdev_ieee802154_submac: set bytes_tx correctly
bytes_tx should not be set to 0 at this point. It should be set to the
length of the package, if there is an error after transmission it will
be set to an error value in submac_tx_done. The value will be evaluated
after a TX_DONE Event, so the pending state 0 that is returned from the
_send function while async sending, is not relevant for bytes_tx.