tests/periph/selftest_shield: Add `NONSTRING` attribute to test data
The test data is not intended to be a zero terminated C string. Just
a bunch of bytes that are easy to trace with a logic analyzer, in
case someone as to debug a test failure.

(cherry picked from commit d96a5a38b5acf538187d43d127d53f79aa995069)