cpu/esp_common: fix FreeRTOS rmutex handling
The ESP-IDF WiFi interface wrapper always calls `xSemaphoreTakeRecursive` and `xSemaphoreGiveRecursive` for mutexes, regardless of whether they are recursive or not. Instead of an `assert`, the fix now also checks the type of the mutex in the functions for recursive mutexes and then calls the functions for normal mutexes if the mutex is not recursive.

(cherry picked from commit 323d26362927167b5ef83eb1b4a7d9b44b43ec09)